chore: decouple `Filter` between `core` and `sdk` moves `SubscriptionManager` to `sdk` side
